SF: Add support for 64 bit consumer usage flag
Change consumer usage query calls to get 64 bit flags.
Make modifications to accomodate 64 bit flags.

CRs-Fixed: 2223978

+++ services/surfaceflinger/ExSurfaceFlinger/ExVirtualDisplaySurface.h
@@ -46,7 +46,7 @@ private:
     virtual status_t beginFrame(bool mustRecompose);
     virtual void setOutputUsage(uint64_t flag);
     bool mSecure;
-    int mSinkUsage;
+    uint64_t mSinkUsage;
 };
 
 }; // namespace android
